- Agent prompt for basic page creation - Design creation and link switched to direct API call
